# openapi.yaml
openapi: 3.0.2
info:
  title: Catalog history service
  description: desc
  version: '1.0'
  termsOfService: 'https://fellesdatakatalog.digdir.no/about'
  contact:
    name: Digitaliseringsdirektoratet
    url: 'https://fellesdatakatalog.digdir.no'
    email: fellesdatakatalog@digdir.no
  license:
    name: License of API
    url: 'http://data.norge.no/nlod/no/2.0'
externalDocs:
  description: Dokumentasjon på Informasjonsforvaltning GitHub
  url: https://informasjonsforvaltning.github.io/
tags:
  - name: catalog-history-service
    description: Catalog history service
servers:
  - url: 'https://catalog-history.fellesdatakatalog.digdir.no'
paths:
  /concepts/{conceptId}/updates:
    post:
      tags:
        - catalog-history-service
      summary: Add update to concept
      description: Adds update to concept history
      operationId: update
      parameters:
        - name: conceptId
          in: path
          description: concept id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: No content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePayload'
    get:
      tags:
        - catalog-history-service
      summary: Query for update history.
      description: Returns a collection of updates
      operationId: getHistory
      parameters:
        - name: conceptId
          in: path
          description: concept id
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: page number
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: page size
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          description: sort type
          schema:
            type: string
            enum:
              - date_desc
              - date_asc
            default: date_desc
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Updates'
  /concepts/{conceptId}/updates/{updateId}:
    get:
      tags:
        - catalog-history-service
      description: Get update diff
      operationId: getUpdateDiff
      parameters:
        - name: conceptId
          in: path
          description: concept id
          required: true
          schema:
            type: string
        - name: updateId
          in: path
          description: update id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Update'

components:
  schemas:
    UpdatePayload:
      type: object
      description: Json patch update and meta data regarding update
      properties:
        person:
          $ref: "#/components/schemas/Person"
          description: Person who updated the document        
        operations:
          type: array
          items:
            $ref: '#/components/schemas/JsonPatchOperation'
          description: List of operations defining the update

    Update:
      type: object
      description: Complete data object regarding update
      properties:
        id:
          type: string
          description: update id
        resourceId:
          type: string
          description: resource id
        person:
          $ref: "#/components/schemas/Person"
          description: Person who updated the document 
        datetime:
          type: string
          format: date
          description: date of update
        operations:
          type: array
          items:
            $ref: '#/components/schemas/JsonPatchOperation'
          description: List of operations defining the update

    Person:
      type: object
      description:
        Person
      properties:
        id:
          type: string
          description: sso id
        email:
          type: string
          description: email
        name:
          type: string
          description: full name

    JsonPatchOperation:
      type: object
      description: 'A JSON patch document as defined by RFC 6902'
      required:
        - op
        - path
      properties:
        op:
          type: string
          nullable: false
          description: 'The operation to be performed'
          enum:
            - add
            - remove
            - replace
            - move
            - copy
            - test
        path:
          type: string
          nullable: false
          description: 'A JSON pointer to the property being affected'
        value:
          type: object
          description: 'The value to add, replace or test.'
        from:
          description: 'A JSON Pointer path pointing to the location to move/copy from.'
          type: string

    Updates:
      type: object
      description: List of updates with meta data
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/Update'
          description: List of updates
        page:
          $ref: "#/components/schemas/Pagination"
          description: Pagination details

    Pagination:
      type: object
      description: Pagination details
      properties:
        currentPage:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer